+++ package.json
@@ -41,25 +41,15 @@
   "dependencies": {
     "@fasteerjs/exceptions": "^1.0.2",
     "@fasteerjs/fasteer": "^3.1.1",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@sentry/node": "^6.3.4",
     "@sysdotini/erela.js-spotify": "^1.2.4",
     "axios": "^0.21.1",
-    "bulma": "^0.9.2",
-    "bulma-checkradio": "^1.1.1",
-    "bulmaselect": "^2.0.8",
     "canvas": "^2.7.0",
     "chalk": "^4.1.1",
     "cookie-parser": "^1.4.5",
-    "cors": "^2.8.5",
-    "csurf": "^1.11.0",
     "dayjs": "^1.10.4",
     "erela.js": "^2.3.3",
     "eris": "^0.15.0",
-    "express": "^4.17.1",
-    "express-rate-limit": "^5.2.6",
-    "express-session": "^1.17.1",
     "fastify": "^3.17.0",
     "fastify-accepts": "^2.0.1",
     "fastify-cookie": "^5.3.1",
@@ -69,10 +59,8 @@
     "fastify-secure-session": "^2.3.1",
     "fastify-session": "^5.2.1",
     "fastify-static": "^4.2.2",
-    "helmet": "^4.5.0",
     "ioredis": "^4.27.6",
     "liquidjs": "^9.23.4",
-    "passport": "^0.4.1",
     "passport-discord": "^0.1.4",
     "point-of-view": "^4.14.0",
     "rethinkdb-ts": "^2.4.10",
@@ -85,12 +73,6 @@
   "devDependencies": {
     "@sentry/types": "^6.3.5",
     "@types/accepts": "^1.3.5",
-    "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.10",
-    "@types/csurf": "^1.11.1",
-    "@types/express": "4.17.11",
-    "@types/express-rate-limit": "^5.1.1",
-    "@types/express-session": "^1.17.3",
     "@types/ioredis": "^4.26.4",
     "@types/node": "^15.0.1",
     "@types/passport-discord": "^0.1.3",
@@ -106,7 +88,6 @@
     "nodemon": "^2.0.7",
     "pm2": "^4.5.6",
     "prettier": "^2.2.1",
-    "purgecss": "^4.0.3",
     "rimraf": "^3.0.2",
     "sass": "^1.32.12",
     "ts-node": "^9.1.1",
