--- src/_webserver/lifecycle/middleware/auth.ts
@@ -1,8 +1,0 @@
-import type { RouteDef } from "_webserver/types";
-import { onlyRoutes } from "_webserver/utils";
-
-const ONLY_ROUTES: RouteDef[] = [];
-
-export const authMiddleware = onlyRoutes(async (req, res) => {
-  // TODO: implement auth
-}, ONLY_ROUTES);
