+++ src/webserver/index.ts
@@ -1,126 +1,66 @@
-/**
- * @file Webserver
- * @description Webserver for dashboard and voting
- * @module webserver/index
- */
-
-import type { NextFunction, Request, Response } from "express";
-import type { SessionOptions } from "express-session";
-import type { HibikiClient } from "../classes/Client";
+import { createFasteer } from "@fasteerjs/fasteer";
+import type { HibikiClient } from "classes/Client";
 import { Liquid } from "liquidjs";
-import { RethinkDBStore } from "session-rethinkdb-ts";
-import { readFileSync, readdirSync } from "fs";
-import { minify } from "terser";
-import { apiRoutes } from "./routes/api";
-import { authRoutes } from "./routes/auth";
-import { indexRoutes } from "./routes/index";
-import { manageRoutes } from "./routes/manage";
-import { votingRoutes } from "./routes/voting";
-import { loadIcons } from "../utils/webserver";
-import cookieParser from "cookie-parser";
-import cors from "cors";
-import crypto from "crypto";
-import csurf from "csurf";
-import express from "express";
-import expressSession from "express-session";
-import helmet from "helmet";
-import passport from "passport";
+import { fastify } from "fastify";
+import { lifecycle } from "./lifecycle";
+import fastifyAccepts from "fastify-accepts";
+import fastifyCsrf from "fastify-csrf";
 import path from "path";
+import pointOfView from "point-of-view";
+import fastifyStatic from "fastify-static";
+import helmet from "helmet";
+import { sessions } from "./lifecycle/sessions";
+
+// The old express webserver. Temporary till this new one is done.
+const OLD_WEBSERVER = path.join(__dirname, "..", "webserver");
 
 const isProduction = process.env.NODE_ENV === "production";
-const LAYOUTS_DIRECTORY = path.join(__dirname, "layouts");
-const PARTIALS_DIRECTORY = path.join(__dirname, "partials");
-const PUBLIC_DIRECTORY = path.join(__dirname, "public");
-const VIEWS_DIRECTORY = path.join(__dirname, "views");
-const cookieSecret = crypto.randomBytes(48).toString("hex");
+const LAYOUTS_DIRECTORY = path.join(OLD_WEBSERVER, "layouts");
+const PARTIALS_DIRECTORY = path.join(OLD_WEBSERVER, "partials");
+const PUBLIC_DIRECTORY = path.join(OLD_WEBSERVER, "frontend", "dist");
+const VIEWS_DIRECTORY = path.join(OLD_WEBSERVER, "views");
+const CONTROLLERS_DIRECTORY = path.join(__dirname, "controllers", "**", "*Controller.{ts,js}");
+const COOKIE_SECRET_PATH = path.join(__dirname, "..", "..", "hibiki_cookie_secret.key");
 
-export function startWebserver(bot: HibikiClient) {
-  const app = express();
-  app.enable("trust proxy");
-  app.disable("x-powered-by");
-
-  // Opts out of Chrome FLoC, even though we don't serve ads. https://amifloced.org/
-  app.use((_req, res, next) => {
-    res.header("Permissions-Policy", "interest-cohort=()");
-    next();
-  });
-
-  // Disables cache-control on specified routes
-  const noCache = (_req: Request, res: Response, next: NextFunction) => {
-    res.header("Cache-Control", "no-cache");
-    next();
-  };
-
-  // Enables CSRF protection on specified routes
-  const csurfProtection = csurf({ cookie: true });
-
-  // RethinkDB session store
-  const sessionStore = new RethinkDBStore({
-    connectOptions: {
-      db: bot.config.database.db,
-      password: bot.config.database.password,
-      port: bot.config.database.port,
-      host: bot.config.database.host,
-      user: bot.config.database.user,
-      silent: true,
+export const startWebserver = async (bot: HibikiClient) => {
+  const app = createFasteer(
+    {
+      controllers: CONTROLLERS_DIRECTORY,
+      port: bot.config.dashboard.port,
+      host: bot.config.dashboard.host,
+      helmet: {
+        hidePoweredBy: true,
+        // @ts-ignore - there are only `helmet` types included, however this is a valid fastify-helmet option.
+        // See fasteerjs/fasteer#4.
+        enableCSPNonces: true,
+        contentSecurityPolicy: {
+          directives: {
+            ...helmet.contentSecurityPolicy.getDefaultDirectives(),
+            "img-src": ["'self'", "cdn.discordapp.com", "data:"],
+            "script-src": ["'self'"],
+          },
+        },
+      },
+      cors: {
+        credentials: true,
+      },
     },
-  });
-
-  // Express session options
-  const sessionOptions: SessionOptions = {
-    secret: cookieSecret,
-    store: sessionStore,
-    name: "Session",
-    resave: false,
-    saveUninitialized: false,
-    unset: "destroy",
-    cookie: {
-      maxAge: 1000 * 60 * 60 * 24 * 7,
-      signed: true,
-      path: "/",
-      httpOnly: true,
-      sameSite: "lax",
-      // If you are getting infinite redirects, set this to "false" if you don't have a HTTPS-only environment setup.
-      secure: bot.config.dashboard.cookieSecure,
-    },
-  };
-
-  // Sets nonce locale
-  app.use((_req, res, next) => {
-    res.locals.nonce = crypto.randomBytes(16).toString("hex");
-    next();
-  });
-
-  // Enables helmet
-  app.use(
-    helmet({
-      contentSecurityPolicy: false,
-      hidePoweredBy: false,
+    fastify({
+      trustProxy: bot.config.dashboard.trustProxy,
     }),
   );
 
-  // Enables our CSP options
-  app.use((req, res, next) => {
-    helmet.contentSecurityPolicy({
-      directives: {
-        ...helmet.contentSecurityPolicy.getDefaultDirectives(),
-        "img-src": ["'self'", "cdn.discordapp.com"],
-        "script-src": ["'self'", `'nonce-${res.locals.nonce}'`],
-      },
-    })(req, res, next);
-  });
+  app
+    .inject({
+      bot,
+    })
+    .plugin(lifecycle);
 
-  // Middleware
-  app.use(cors({ credentials: true }));
-  app.use(cookieParser(cookieSecret));
-  app.use(expressSession(sessionOptions));
-  app.use(express.urlencoded({ extended: true, limit: "1mb" }));
-  app.use(express.json({ limit: "1mb" }));
-  app.use(passport.initialize());
-  app.use(passport.session());
+  sessions(bot, {
+    secretPath: COOKIE_SECRET_PATH,
+  })(app);
 
-  // Liquid engine options
-  const engine = new Liquid({
+  const liquid = new Liquid({
     root: [VIEWS_DIRECTORY, PARTIALS_DIRECTORY, LAYOUTS_DIRECTORY],
     cache: isProduction,
     lenientIf: true,
@@ -128,66 +68,25 @@ export function startWebserver(bot: HibikiClient) {
     extname: ".liquid",
   });
 
-  // Custom icon tag
-  loadIcons(engine);
-
-  // View engine options
-  app.engine("liquid", engine.express());
-  app.set("views", [VIEWS_DIRECTORY, PARTIALS_DIRECTORY, LAYOUTS_DIRECTORY]);
-  app.set("view engine", "liquid");
-
-  // Minifies JS if running in production
-  if (process.env.NODE_ENV === "production") {
-    const files = readdirSync(`${PUBLIC_DIRECTORY}/js`, { withFileTypes: true, encoding: "utf-8" });
-    files.forEach(async (file) => {
-      if (file.isDirectory()) return;
-      if (!file.name.endsWith(".js")) return;
-      const fileSource = readFileSync(`${PUBLIC_DIRECTORY}/js/${file.name}`, { encoding: "utf-8" });
-      const minifiedFile = await minify(fileSource, {
-        sourceMap: false,
-        mangle: true,
-        format: {
-          indent_level: 2,
-          comments: false,
-        },
-        compress: {
-          defaults: true,
-        },
-      });
-
-      // Falls back to using non-minified files
-      if (!minifiedFile.code) {
-        bot.log.error(`Error while minifying ${file.name}, the non-minified one will be served instead.`);
-        app.use(`${PUBLIC_DIRECTORY}/js/${file.name}`, (_req, res) => {
-          return res.send(fileSource);
-        });
-      }
-
-      // Uses the minified files
-      app.use(`${PUBLIC_DIRECTORY}/js/${file.name}`, (_req, res) => {
-        res.set("Content-Type", "application/javascript").send(minifiedFile.code);
-      });
+  app.fastify
+    .register(fastifyAccepts)
+    .register(fastifyCsrf)
+    .register(pointOfView, {
+      engine: {
+        liquid,
+      },
+      viewExt: "liquid",
+      root: VIEWS_DIRECTORY,
+      defaultContext: {
+        bot,
+        botAvatar: bot.user.dynamicAvatarURL(null, 512),
+      },
+    })
+    .register(fastifyStatic, {
+      root: PUBLIC_DIRECTORY,
+      prefix: "/",
     });
-  }
 
-  // Uses public folder
-  app.use("/public/", express.static(PUBLIC_DIRECTORY));
-
-  // Routes
-  app.use("/vote/", votingRoutes(bot));
-  app.use("/api/", csurfProtection, apiRoutes(bot));
-  app.use("/auth/", csurfProtection, noCache, authRoutes(bot.user));
-  app.use("/manage/", csurfProtection, noCache, manageRoutes(bot));
-  app.use("/", csurfProtection, indexRoutes(bot));
-
-  // 404 handler
-  app.use((req, res) => {
-    if (req.accepts("html")) return res.status(404).render("404", { url: req.url, errorCode: 404 });
-    else if (req.accepts("json")) return res.status(404).send({ error: "404" });
-    else res.status(404).type("txt").send("404");
-  });
-
-  app.listen(bot.config.dashboard.port, "0.0.0.0", () => {
-    bot.log.info(`Webserver running on port ${bot.config.dashboard.port}`);
-  });
-}
+  const addr = await app.start();
+  bot.log.info(`Webserver running at ${addr}`);
+};
