--- src/webserver/layouts/default.liquid
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    {% include "head" %} {% block head %} {% endblock %}
-  </head>
-
-  <body>
-    <!-- Main content -->
-    {% block content %} {% endblock %}
-
-    <!-- Footer -->
-    {% include "footer" %}
-  </body>
-</html>
