--- src/webserver/layouts/error.liquid
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta content="IE=edge,chrome=1" http-equiv="X-UA-Compatible" />
-    <meta content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0" name="viewport" />
-    <title>{{ error }}</title>
-  </head>
-
-  <body>
-    {% block content %} {% endblock %}
-  </body>
-</html>
