--- src/webserver/partials/buttons.liquid
@@ -1,33 +1,0 @@
-<div class="sticky-buttons">
-  <div class="field has-addons">
-    <div class="buttons">
-      <!-- Back to top -->
-      <p class="control">
-        <a href="#" class="button">
-          <span class="icon is-small fa-xs">{{ "arrow-up" | icon }}</span>
-          <span>{{ locales.web.BACK_TO_TOP }}</span>
-        </a>
-      </p>
-
-      <!-- Save changes -->
-      <p class="control">
-        <a class="button" id="submit">
-          <!-- Handles CSRF validation -->
-          <input type="hidden" id="_csrf" name="_csrf" value="{{ csrfToken }}" />
-          <span class="icon is-small fa-xs">{{ "check" | icon }}</span>
-          <span id="saved">{{ locales.web.SAVE_CHANGES }}</span>
-        </a>
-      </p>
-
-      <!-- Delete config -->
-      <p class="control">
-        <a class="button modal-button" data-target="#modal" aria-haspopup="true">
-          <span class="icon is-small fa-xs">{{ "trash" | icon }}</span>
-          <span>{{ locales.web.RESET_CONFIG }}</span>
-        </a>
-      </p>
-    </div>
-  </div>
-</div>
-
-<br />
