--- src/webserver/partials/footer.liquid
@@ -1,17 +1,0 @@
-<footer class="footer">
-  <div class="content has-text-centered is-size-5">
-    <p>
-      {{ bot.user.username }}, {{ locales.web.BY }} <a href="https://espi.me" target="_blank" rel="noopener noreferrer">sysdotini</a>,
-      <a href="https://github.com/resolvedxd" target="_blank" rel="noopener noreferrer">resolved</a>, {{ locales.web.AND }}
-      <a href="https://github.com/sysdotini/hibiki/graphs/contributors">{{ locales.web.CONTRIBUTORS }}</a>
-      <br />
-      {{ locales.web.SOURCE }} <a href="https://github.com/sysdotini/hibiki/blob/main/LICENSE">{{ locales.web.LICENSED }}</a>
-      {{ locales.web.SOURCE_LICENSE }}
-      <br />
-      <a href="/privacy/" target="_blank" rel="noopener noreferrer">
-        <span class="icon is-small mr-1 fa-lg">{{ "shield-alt" | icon }}</span>
-        <span>{{ locales.web.PRIVACY_POLICY }}</span>
-      </a>
-    </p>
-  </div>
-</footer>
