--- src/webserver/partials/head.liquid
@@ -1,31 +1,0 @@
-<!-- Metadata -->
-<meta charset="utf-8" />
-<meta content="IE=edge,chrome=1" http-equiv="X-UA-Compatible" />
-<meta content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0" name="viewport" />
-<meta content="{{ bot.config.colors.general }}" name="theme-color" />
-<meta content="{{ bot.config.colors.general }}" name="msapplication-navbutton-color" />
-<meta content="{{ bot.config.colors.general }}" name="apple-mobile-web-app-status-bar-style" />
-
-<!-- Resources -->
-<link rel="preconnect" href="https://cdn.discordapp.com" crossorigin="anonymous" />
-<link rel="icon" href="{{ botAvatar }}?size=128" type="image/png" />
-<link rel="stylesheet" href="/public/css/theme.css" type="text/css" />
-
-<!-- Metadata -->
-{% if metaDescription %}
-<meta content="website" property="og:type" />
-<meta content="summary" name="twitter:card" />
-<meta content="@hibikiapp" name="twitter:site" />
-<meta content="@hibikiapp" name="twitter:creator" />
-<meta content="hibiki, hibikiapp, hibiki app, hibiki bot, hibikibot, hibiki discord, hibiki discord bot, discord bot" name="keywords" />
-<meta content="{{ botAvatar }}?size=128" property="og:image" />
-<meta content="{{ botAvatar }}?size=128" name="twitter:image" />
-<meta content="https://hibiki.app" property="og:url" />
-<meta content="{{ bot.user.username }}" property="og:title" />
-<meta content="{{ bot.user.username }}" name="twitter:title" />
-<meta content="{{ metaDescription }}" name="description" />
-<meta content="{{ metaDescription }}" property="og:description" />
-<meta content="{{ metaDescription }}" name="twitter:description" />
-{% endif %}
-
-<title>{{ pageTitle }}</title>
