--- src/webserver/partials/modal.liquid
@@ -1,26 +1,0 @@
-<div class="modal" id="modal">
-  <div class="modal-background"></div>
-  <div class="modal-content">
-    <div class="modal-card">
-      <header class="modal-card-head">
-        {% if page == "profile" %}
-        <p class="modal-card-title">{{ locales.web.DELETE_PROFILE_DATA }}</p>
-        {% else %}
-        <p class="modal-card-title">{{ locales.web.DELETE_SERVER_CONFIG }}</p>
-        {% endif %}
-        <button id="modal-close" class="modal-close is-large delete" aria-label="close"></button>
-      </header>
-
-      <footer class="modal-card-foot">
-        <button class="button is-danger" id="delete">
-          <span class="icon is-small fa-xs">{{ "trash" | icon }}</span>
-          <span>{{ locales.web.RESET }}</span>
-        </button>
-        <button class="button cancel" id="cancel-button">
-          <span class="icon is-small fa-xs">{{ "times-circle" | icon }}</span>
-          <span>{{ locales.web.CANCEL }}</span>
-        </button>
-      </footer>
-    </div>
-  </div>
-</div>
