--- src/webserver/partials/navbar.liquid
@@ -1,88 +1,0 @@
-<div class="hero-head">
-  <nav class="navbar is-transparent" role="navigation" aria-label="main navigation">
-    <div class="container">
-      <div class="navbar-brand">
-        <a href="/" class="navbar-item has-text-white">
-          <img src="{{ botAvatar }}?size=64" class="is-rounded" alt="{{ bot.user.username }}" width="28" height="28" />
-          <span class="is-inline-block ml-2">{{ bot.user.username }}</span>
-        </a>
-        <a class="navbar-burger burger" aria-label="menu" aria-expanded="false" data-target="navMenu" role="button">
-          <span aria-hidden="true" class="has-text-white"></span>
-          <span aria-hidden="true" class="has-text-white"></span>
-          <span aria-hidden="true" class="has-text-white"></span>
-        </a>
-      </div>
-      <div class="navbar-menu" id="navMenu">
-        <div class="navbar-start">
-          <!-- Index page -->
-          {% if page == "/" %}
-
-          <!-- If the user is logged in -->
-          {% if user %}
-          <a href="/manage/servers" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "wrench" | icon }}</span>
-            <span>{{ locales.web.DASHBOARD }}</span>
-          </a>
-          {% endif %}
-
-          <!-- Useful links -->
-          <a href="/donate/" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "heart" | icon }}</span>
-            <span>{{ locales.web.DONATE }}</span>
-          </a>
-          <a href="/support/" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "discord" | icon }}</span>
-            <span>{{ locales.web.SUPPORT }}</span>
-          </a>
-          <a href="/translate/" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "globe" | icon }}</span>
-            <span>{{ locales.web.TRANSLATE }}</span>
-          </a>
-
-          <!-- Servers page -->
-          {% elsif page == "servers" %}
-          <a href="/auth/" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "sync" | icon }}</span>
-            <span>{{ locales.web.REFRESH }}</span>
-          </a>
-
-          <!-- Manage page -->
-          {% elsif page == "manage" %}
-          <a href="/manage/servers" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "home" | icon }}</span>
-            <span>{{ locales.web.SERVERS }}</span>
-          </a>
-
-          <!-- Profile page -->
-          {% elsif page == "profile" %}
-          <a href="/manage/servers" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "home" | icon }}</span>
-            <span>{{ locales.web.SERVERS }}</span>
-          </a>
-          {% endif %}
-        </div>
-
-        <!-- If the user is logged in  -->
-        {% if user %}
-        <div class="navbar-end">
-          <a href="/manage/profile" class="navbar-item has-text-white">
-            <img src="{{ user.avatar }}" class="is-rounded is-inline-block" alt="{{ user.username }}" />
-            <span class="is-inline-block ml-2">{{ user.username }}#{{ user.discriminator }}</span>
-          </a>
-          <a href="/auth/logout/" class="navbar-item has-text-white">
-            <span class="icon is-small mr-1">{{ "sign-out-alt" | icon }}</span>
-            <span>{{ locales.web.LOGOUT }}</span>
-          </a>
-          {% else %}
-          <div class="navbar-end">
-            <a href="/manage/servers/" class="navbar-item has-text-white">
-              <span class="icon is-small mr-1">{{ "sign-in-alt" | icon }}</span>
-              <span>{{ locales.web.LOGIN }}</span>
-            </a>
-          </div>
-          {% endif %}
-        </div>
-      </div>
-    </div>
-  </nav>
-</div>
