--- src/webserver/partials/sidebar.liquid
@@ -1,27 +1,0 @@
-<p class="menu-label is-active">{{ locales.web.SETTINGS }}</p>
-<ul id="menu" class="menu-list">
-  <li>
-    <a href="#general">{{ locales.web.SIDEBAR_GENERAL }}</a>
-  </li>
-  <li>
-    <a href="#roles">{{ locales.web.SIDEBAR_ROLES }}</a>
-  </li>
-  <li>
-    <a href="#greeting">{{ locales.web.SIDEBAR_GREETING }}</a>
-  </li>
-  <li>
-    <a href="#logging">{{ locales.web.SIDEBAR_LOGGING }}</a>
-  </li>
-  <li>
-    <a href="#music">{{ locales.web.SIDEBAR_MUSIC }}</a>
-  </li>
-  <li>
-    <a href="#pinboard">{{ locales.web.SIDEBAR_PINBOARD }}</a>
-  </li>
-  <li>
-    <a href="#sniping">{{ locales.web.SIDEBAR_SNIPING }}</a>
-  </li>
-  <li>
-    <a href="#automod">{{ locales.web.SIDEBAR_AUTOMOD }}</a>
-  </li>
-</ul>
