--- src/webserver/public/scss/custom.scss
@@ -1,169 +1,0 @@
-// Branding
-$hibiki-primary: #8e94f2;
-$hibiki-secondary: #9fa0ff;
-$hibiki-tertiary: #bbadff;
-$hibiki-pink: #dab6fc;
-$hibiki-dark: #17181c;
-$hibiki-primary-darker: #676cb8;
-$hibiki-secondary-darker: #886fb9;
-$hibiki-link: #60a5ff;
-
-.icon {
-  align-items: center;
-  display: inline-flex;
-  justify-content: center;
-  height: 1.75rem;
-  width: 1.75rem;
-}
-
-.menu {
-  z-index: 1;
-}
-
-.emoji-input {
-  width: 3em !important;
-}
-
-.is-checkradio[type="radio"] + label,
-.is-checkradio[type="checkbox"] + label {
-  padding-top: 0 !important;
-  padding-bottom: 0 !important;
-}
-
-.fa-xs {
-  transform: scale(0.75);
-}
-.fa-sm {
-  transform: scale(0.875);
-}
-
-.fa-lg {
-  transform: scale(1.33);
-}
-.fa-2x {
-  transform: scale(2);
-}
-.fa-3x {
-  transform: scale(3);
-}
-
-#logo {
-  box-shadow: 0px 0px 20px $hibiki-tertiary;
-  border-radius: 50%;
-}
-
-.is-rounded {
-  border-radius: 50%;
-}
-
-.navbar-menu {
-  .is-active,
-  .is-inline-block {
-    vertical-align: middle;
-    justify-content: center;
-  }
-}
-
-#stare {
-  width: 15%;
-  height: 15%;
-  top: 100vh;
-  z-index: 0;
-  bottom: 0;
-  position: absolute;
-  height: auto;
-}
-
-a.navbar-item,
-.navbar-toggle,
-.navbar-burger:hover,
-.navbar-menu.is-active,
-.navbar-menu.is-active:hover {
-  background: transparent !important;
-}
-
-.media {
-  align-items: center;
-  justify-content: center;
-  margin-left: auto;
-  margin-right: auto;
-  padding-left: 0;
-}
-
-.sticky {
-  position: sticky;
-  position: -webkit-sticky;
-  top: 30px;
-}
-
-.sticky-buttons {
-  background: #fff;
-  position: sticky;
-  position: -webkit-sticky;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  top: 0px;
-  z-index: 30;
-}
-
-@media (prefers-color-scheme: dark) {
-  .sticky-buttons {
-    background: #17181c;
-  }
-
-  .modal-card {
-    outline: 3px rgba(255, 255, 255, 0.153) solid;
-  }
-}
-
-// Mixin for animations
-@mixin keyframes($name) {
-  @-webkit-keyframes #{$name} {
-    @content;
-  }
-  @-moz-keyframes #{$name} {
-    @content;
-  }
-  @-ms-keyframes #{$name} {
-    @content;
-  }
-  @keyframes #{$name} {
-    @content;
-  }
-}
-
-.gradient {
-  animation: gradient 4s ease infinite;
-  background-size: 400% 400%;
-
-  @include keyframes(gradient) {
-    0% {
-      background: $hibiki-primary-darker;
-    }
-
-    50% {
-      background: $hibiki-secondary-darker;
-    }
-
-    100% {
-      background: $hibiki-primary-darker;
-    }
-  }
-}
-
-// Custom floating
-.floating {
-  animation: float 2s infinite ease-in-out;
-
-  @include keyframes(float) {
-    0% {
-      transform: translateY(0);
-    }
-    50% {
-      transform: translateY(-5px);
-    }
-    100% {
-      transform: translateY(0px);
-    }
-  }
-}
