--- src/webserver/public/scss/dark.scss
@@ -1,59 +1,0 @@
-@charset "utf-8";
-
-// Branding
-$hibiki-primary: #8e94f2;
-$hibiki-secondary: #9fa0ff;
-$hibiki-tertiary: #bbadff;
-$hibiki-pink: #dab6fc;
-$hibiki-dark: #17181c;
-$hibiki-primary-darker: #676cb8;
-$hibiki-secondary-darker: #886fb9;
-$hibiki-link: #60a5ff;
-$hibiki-link-hover: #7db5ff;
-
-// Inverted default Bulma colors
-$black-inverted: #f5f5f5;
-$black-bis-inverted: #ededed;
-$black-ter-inverted: #dbdbdb;
-$grey-darker-inverted: #c9c9c9;
-$grey-dark-inverted: #b5b5b5;
-$grey-inverted: #858585;
-$grey-light-inverted: #4a4a4a;
-$grey-lighter-inverted: #242424;
-$grey-lightest-inverted: #121212;
-$white-ter-inverted: #0a0a0a;
-$white-bis-inverted: #050505;
-$white-inverted: #000000;
-
-// Elements to invert
-$grey-darker: $grey-darker-inverted;
-$grey-dark: $grey-dark-inverted;
-$grey: $grey-inverted;
-$grey-light: $grey-light-inverted;
-$grey-lighter: $grey-lighter-inverted;
-$grey-lightest: $grey-lightest-inverted;
-$white-ter: $white-ter-inverted;
-$white-bis: $white-bis-inverted;
-
-// Color overrides
-$turquoise: $hibiki-primary;
-$blue: $hibiki-link;
-$border: $grey-light-inverted;
-$border-light: $grey-lighter-inverted;
-$link-hover-border: $grey-inverted;
-$border-hover: $grey-light-inverted;
-$link-hover: $hibiki-link-hover;
-
-// Cleanup
-$light: #f5f5f5;
-$scheme-main: $white-bis;
-$scheme-main-bis: #0d0d0f;
-$body-background-color: #17181c;
-
-// Bulma-checkradio overrides
-$checkbox-border: 0.1rem solid $grey-darker-inverted;
-$checkradio-focus: 1px dotted $grey-dark-inverted;
-$checkbox-block-background: $black-ter-inverted;
-
-// yea it is
-@import "./default.scss";
