--- src/webserver/public/scss/default.scss
@@ -1,22 +1,0 @@
-// Branding
-$hibiki-primary: #8e94f2;
-$hibiki-secondary: #9fa0ff;
-$hibiki-tertiary: #bbadff;
-$hibiki-pink: #dab6fc;
-$hibiki-dark: #17181c;
-$hibiki-primary-darker: #676cb8;
-$hibiki-secondary-darker: #886fb9;
-$hibiki-link: #3273dc;
-$hibiki-primary-darker: #676cb8;
-$hibiki-secondary-darker: #886fb9;
-
-// Color overrides
-$turquoise: $hibiki-primary !default;
-$cyan: #3273dc !default;
-$body-background-color: #fff !default;
-$footer-padding: 3em;
-
-@import "./custom.scss";
-@import "../../../../node_modules/bulma/bulma.sass";
-@import "../../../../node_modules/bulmaselect/src/scss/bulmaselect.scss";
-@import "../../../../node_modules/bulma-checkradio/src/sass/index.sass";
