--- src/webserver/public/scss/theme.scss
@@ -1,3 +1,0 @@
-@charset "utf-8";
-@import "../css/dark.css" (prefers-color-scheme: dark);
-@import "../css/default.css" (prefers-color-scheme: light);
