+++ src/webserver/types/fastify.d.ts
