--- src/webserver/views/404.liquid
@@ -1,14 +1,0 @@
-<!-- Page title -->
-{% assign error = "404" %}
-
-<!-- Default layout -->
-{% layout "error" %}
-
-<!-- Error content -->
-{% block content %}
-
-<div>
-  <p>404 Not Found</p>
-</div>
-
-{% endblock %}
