--- src/webserver/views/authfail.liquid
@@ -1,14 +1,0 @@
-<!-- Page title -->
-{% assign error = "Error" %}
-
-<!-- Default layout -->
-{% layout "error" %}
-
-<!-- Error content -->
-{% block content %}
-
-<div>
-  <p>Failed to authenticate.</p>
-</div>
-
-{% endblock %}
