--- src/webserver/views/index.liquid
@@ -1,63 +1,0 @@
-<!-- Page title -->
-{% assign pageTitle = bot.user.username %}
-
-<!-- Metadata tagline -->
-{% assign metaDescription = "The best all-in-one Discord bot! Automod, fun, music, utilities, and more." %}
-
-<!-- Default layout -->
-{% layout "default" %}
-
-<!-- Begin content  -->
-{% block content %}
-
-<section class="hero gradient is-fullheight">
-  {% include "navbar" %}
-  <div class="hero-body">
-    <div class="container has-text-centered animate__animated animate__fadeIn animate__faster">
-      <img src="{{ botAvatar }}" id="logo" alt="{{ bot.user.username }}>" width="300" height="300" />
-      <p class="is-size-2 has-text-weight-semibold has-text-light">{{ bot.user.username }}</p>
-      <p class="control has-text-centered">
-        <a href="/invite/" class="button is-large is-rounded is-light is-size-4-desktop is-size-5-mobile">
-          <span class="icon is-small fa-xs">{{ "plus" | icon }}</span>
-          <span>{{ locales.web.INVITE }} {{ bot.user.username }}</span>
-        </a>
-      </p>
-    </div>
-  </div>
-
-  <div class="hero-footer animate__animated animate__fadeInUp animate__faster">
-    <div class="tabs is-centered is-fullwidth">
-      <div class="container">
-        <ul>
-          <li>
-            <a href="#stats" class="has-text-light">
-              <span class="icon is-small"> {{ "chart-bar" | icon }} </span>
-              <p>{{ locales.web.STATISTICS }}</p>
-            </a>
-          </li>
-          <li>
-            <a href="#features" class="has-text-light">
-              <span class="icon is-small">{{ "magic" | icon }}</span>
-              <p>{{ locales.web.FEATURES }}</p>
-            </a>
-          </li>
-          <li>
-            <a href="#about" class="has-text-light">
-              <span class="icon is-small">{{ "info-circle" | icon }}</span>
-              <p>{{ locales.web.ABOUT }}</p>
-            </a>
-          </li>
-        </ul>
-      </div>
-    </div>
-  </div>
-</section>
-
-<!-- About -->
-{% include "about" %}
-
-<!-- Scripts -->
-<script src="/public/js/navbar.js" nonce="{{ nonce }}"></script>
-
-<!-- End content  -->
-{% endblock %}
