--- src/webserver/views/profile.liquid
@@ -1,166 +1,0 @@
-<!-- Page title -->
-{% assign pageTitle = locales.web.PROFILE %}
-
-<!-- Default layout -->
-{% layout "default" %}
-
-<!-- Additional head data -->
-{% block head %}
-<meta name="user-id" content="{{ user.id }}" />
-<meta name="csrf-token" content="{{ csrfToken }}" />
-<meta name="save-changes" content="{{ locales.web.SAVE_CHANGES }}" />
-<meta name="changes-saved" content="{{ locales.web.CHANGES_SAVED }}" />
-{% endblock %}
-
-<!-- Start content -->
-{% block content %}
-
-<section class="hero gradient">
-  {% include "navbar" %}
-  <div class="hero-body">
-    <div class="container">
-      <p class="title has-text-white is-size-3">{{ locales.web.MANAGE_PROFILE_WELCOME }}, {{ user.username }}.</p>
-      <p class="subtitle has-text-white is-size-5">{{ locales.web.MANAGE_PROFILE_DESCRIPTION }}</p>
-    </div>
-  </div>
-</section>
-
-<section class="section container">
-  {% include "buttons" %}
-
-  <!-- Cycles through each item -->
-  {% for item in items %}
-  <div class="field">
-    <p class="title is-size-4">{{ itemLocalesTitle[item.id] }}</p>
-    <p class="subtitle is-size-5">{{ itemLocales[item.id] }}</p>
-
-    {% case item.type %}
-
-    <!-- Strings -->
-    {% when "string" %}
-    <div class="control">
-      <!-- If the item is set -->
-      {% if cfg[item.id] %}
-      <input
-        class="input"
-        id="{{ item.id }}"
-        type="text"
-        value="{{ cfg[item.id] }}"
-        placeholder="{{ item.maximum }}{% if item.maximum > 15 %} {{ locales.web.MANAGE_INPUT_REDUCE }} {% else %} {{
-          locales.web.MANAGE_INPUT_MAXIMUM
-        }} {% endif %}"
-      />
-
-      {% else %}
-      <input
-        class="input"
-        id="{{ item.id }}"
-        type="text"
-        placeholder="{{ item.maximum }}{% if item.maximum > 15 %} {{ locales.web.MANAGE_INPUT_REDUCE }} {% else %} {{
-          locales.web.MANAGE_INPUT_MAXIMUM
-        }} {% endif %}"
-      />
-
-      {% endif %}
-    </div>
-
-    {% else %}
-
-    <!-- Pronouns -->
-    {% when "pronouns" %}
-    <div class="control">
-      <div id="{{ item.id }}" class="select">
-        <select>
-          {% assign pronounsize = pronouns.size | minus:1 %}
-          <!-- Cycles through each pronoun -->
-          {% for i in (0..pronounsize) %}
-          <!-- Sets it as selected -->
-          {% assign isSelected = cfg[item.id] == i %}
-          <!-- If isSelected is set -->
-          {% if isSelected == true %}
-          <!-- There are two "selected" ? -->
-          <option selected>{{ locales.global[pronouns[i]] }}</option>
-          {% else %}
-          <!-- Lists the pronouns -->
-          <option>{{ locales.global[pronouns[i]] }}</option>
-          {% endif %} {% endfor %}
-        </select>
-      </div>
-    </div>
-
-    {% else %}
-
-    <!-- Timezone -->
-    {% when "timezone" %}
-    <div class="control">
-      <input disabled class="input" id="{{ item.id }}" type="text" placeholder="{{ locales.web.MANAGE_TIMEZONE_ERROR }}" />
-    </div>
-
-    {% else %}
-
-    <!-- Locales -->
-    {% when "locale" %}
-    <div class="control">
-      <div id="{{ item.id }}" class="select">
-        <select>
-          {% for locale in localesKeys %}
-          <!-- Checks to see if the locale is set -->
-          {% if cfg[item.id] %} {% assign isSelected = cfg[item.id] == locale %}
-          <!-- Checks to see if the locale is the default -->
-          {% else %} {% assign isSelected = locale == item.default %} {% endif %}
-          <!-- If the locale is selected -->
-          {% if isSelected == true %}
-          <option selected id="{{ locale }}">{{ localeNames[locale}}</option>
-          <!-- Lists each locale -->
-          {% else %}
-          <option id="{{ locale }}">{{ localeNames[locale] }}</option>
-          {% endif %} {% endfor %}
-        </select>
-      </div>
-    </div>
-
-    {% else %}
-
-    <!-- Booleans -->
-    {% when "boolean" %}
-
-    <!-- Checks to see if the item is enabled -->
-    {% if cfg[item.id] %}
-    <!-- Marks it as enabled if it's set -->
-    {% assign isSelected = cfg[item.id] %} {% else %}
-    <!-- Uses the default -->
-    {% assign isSelected = item.default or false %} {% endif %}
-    <div class="field" id="{{ item.id }}" class="control">
-      <!-- If the item is enabled -->
-      {% if isSelected == true %}
-      <input checked id="{{ item.id }}Switch" type="checkbox" name="{{ item.id }}Switch" class="is-checkradio switch is-rounded m-0" />
-      <!-- If the item is disabled -->
-      {% else %}
-      <input id="{{ item.id }}Switch" type="checkbox" name="{{ item.id }}Switch" class="is-checkradio switch is-rounded m-0" />
-      {% endif %}
-      <!-- Item label -->
-      <label for="{{ item.id }}Switch" class="m-0">
-        <span>{{ locales.web.ENABLED }}</span>
-      </label>
-    </div>
-
-    {% else %}
-
-    <!-- Numbers -->
-    {% when "number" %}
-    <input type="number" id="{{ item.id }}" min="{{ item.minimum }}" max="{{ item.maximum }}" value="{{ cfg[item.id] }}" />
-
-    {% endcase %}
-    <br />
-  </div>
-
-  {% endfor %}
-</section>
-
-{% include "modal" %}
-
-<script src="/public/js/modal.js" nonce="{{ nonce }}"></script>
-<script src="/public/js/navbar.js" nonce="{{ nonce }}"></script>
-<script src="/public/js/profile.js" nonce="{{ nonce }}"></script>
-
-{% endblock %}
