--- src/webserver/views/ratelimited.liquid
@@ -1,14 +1,0 @@
-<!-- Page title -->
-{% assign error = "429" %}
-
-<!-- Default layout -->
-{% layout "error" %}
-
-<!-- Error content -->
-{% block content %}
-
-<div>
-  <p>Too many requests in a short period of time. Try again later.</p>
-</div>
-
-{% endblock %}
