--- src/webserver/views/servers.liquid
@@ -1,58 +1,0 @@
-<!-- Page title -->
-{% assign pageTitle = locales.web.SERVERS %}
-
-<!-- Default layout -->
-{% layout "default" %}
-
-<!-- Begin content  -->
-{% block content %}
-
-<section class="hero gradient">
-  {% include "navbar" %}
-  <div class="hero-body">
-    <div class="container">
-      <p class="title has-text-white is-size-2">{{ locales.web.MANAGE_SERVERS_WELCOME }}, {{ user.username }}.</p>
-      <p class="subtitle has-text-white is-size-5">{{ locales.web.MANAGE_SERVERS_DESCRIPTION }}</p>
-    </div>
-  </div>
-</section>
-
-<section class="section container">
-  <!-- Renders each guild -->
-  {% for guild in guilds %}
-  <article class="media">
-    <figure class="media-left">
-      <div class="image is-48x48">
-        <img src="{{ guild.icon }}" alt="{{ guild.name }}" class="is-rounded" width="256" height="256" />
-      </div>
-    </figure>
-    <div class="media-content is-clipped">
-      <p class="subtitle is-size-2">{{ guild.name }}</p>
-    </div>
-
-    <div class="media-right">
-      <!-- If the bot is in the guild -->
-      {% if guild.isIn == true %}
-      <a href="/manage/{{ guild.id }}" class="button is-medium">
-        <span class="icon is-small fa-xs">{{ "wrench" | icon }}</span>
-        <p>{{ locales.web.CONFIG }}</p>
-      </a>
-
-      <!-- If the bot isn't in the guild -->
-      {% else %}
-      <a href="/invite?guild={{ guild.id }}" class="button is-medium">
-        <span class="icon is-small fa-xs">{{ "plus" | icon }}</span>
-        <p>{{ locales.web.INVITE }}</p>
-      </a>
-      {% endif %}
-    </div>
-  </article>
-
-  {% endfor %}
-</section>
-
-<!-- Scripts -->
-<script src="/public/js/navbar.js" nonce="{{ nonce }}"></script>
-
-<!-- End content  -->
-{% endblock %}
